feat: implement global refresh service for Twitch tokens and schedule hourly updates
